include run-sw-tests.js via import-scripts message
