+++ serviceworker.js
@@ -25,6 +25,7 @@ l4.importScripts('src/sw/messaging.js');
 l4.importScripts('src/sw/logging.js');
 
 l4.importScripts('src/sw/messaging-tasks/github/callback.js'); // #TODO Refactor #JensLincke
+l4.importScripts('src/sw/messaging-tasks/import-script.js');
 
 console.log('Service Worker: File Start');
